image page/modal improvements
